+++ src/components/ids-data-grid/demos/columns-click-callback.ts
@@ -68,7 +68,7 @@ const dataGrid = document.querySelector<IdsDataGrid>('#data-grid-columns-click-c
     id: 'inStock',
     name: 'In Stock',
     field: 'inStock',
-    sortable: true,
+    sortable: false,
     resizable: true,
     reorderable: true,
     align: 'center',