+++ src/components/ids-global/ids-global.ts
@@ -1,17 +1,19 @@
 /**
  * IDS Global Functions / Instances / And Utils Exposed to the end user.
  */
+import { IdsDeferred } from '../../utils/ids-deferred-utils/ids-deferred-utils';
 import IdsLocale from '../ids-locale/ids-locale';
 
 declare global {
   interface Window {
     IdsGlobal: {
-      locale?: IdsLocale
+      locale?: IdsLocale;
+      themeLoaded?: IdsDeferred;
     }
   }
 }
 
-window.IdsGlobal = {};
+window.IdsGlobal ??= {};
 
 class IdsGlobal {
   static getLocale(): IdsLocale {
@@ -19,6 +21,14 @@ class IdsGlobal {
       window.IdsGlobal.locale = new IdsLocale();
     }
     return window.IdsGlobal.locale;
+  }
+
+  static onThemeLoaded(): IdsDeferred {
+    if (!window.IdsGlobal.themeLoaded) {
+      window.IdsGlobal.themeLoaded = new IdsDeferred();
+    }
+
+    return window.IdsGlobal.themeLoaded;
   }
 }